refactor(tests): split property_safe_fix.rs into ast/config/strategy submodules
tests/property_safe_fix.rs grew to ~890 lines covering: the
named-config matrix, the synthetic YAML AST + renderers, proptest
strategies, and three property invariants + several deterministic
regressions. The first three are reusable building blocks that change
independently of the test invariants; the file is now broken up as:

  tests/property_safe_fix.rs               (231 lines, proptest!
                                            invariants + deterministic
                                            regressions)
  tests/property_safe_fix/config.rs        (171 lines, named-config
                                            matrix + parse/lint helpers)
  tests/property_safe_fix/ast.rs           (320 lines, AST types +
                                            renderers + residue predicate)
  tests/property_safe_fix/strategy.rs      (214 lines, proptest
                                            strategies)

Each submodule has a single responsibility. Cargo's integration-test
discovery only treats `tests/*.rs` (top-level) as test binaries, so the
subdirectory is purely internal to the property_safe_fix binary.

Refs #221
